clean

+++ C00/ex01/inc/PhoneBook.hpp
@@ -1,20 +1,8 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   PhoneBook.hpp                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: cde-la-r <code@cesardelarosa.xyz>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/20 11:44:53 by cde-la-r          #+#    #+#             */
-/*   Updated: 2025/04/20 11:47:42 by cesi             ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
 #pragma once
 #include "Contact.hpp"
+#include <cstddef>
 
-class   PhoneBook
-{
+class PhoneBook {
     public:
         PhoneBook();
         ~PhoneBook();
@@ -23,8 +11,8 @@ class   PhoneBook
         void    search();
 
     private:
-        static const int    _max = 8;
+        static const size_t _max = 8;
         Contact             _contacts[_max];
-        int                 _next;
-        int                 _count;
+        size_t              _next;
+        size_t              _count;
 };
